moved treatment_summary calculation to share with biosample and donors
