+++ src/encoded/types/shared_calculated_properties.py
@@ -1,6 +1,16 @@
 from snovault import calculated_property
 
 
+def pluralize(value, value_units):
+    try:
+        if float(value) == 1:
+            return str(value) + ' ' + value_units
+        else:
+            return str(value) + ' ' + value_units + 's'
+    except:
+        return str(value) + ' ' + value_units + 's'
+
+
 class CalculatedAward:
     @calculated_property(schema={
         "title": "Award",
@@ -102,3 +112,49 @@ class CalculatedBiosampleSummary:
             elif bs_obj.get('summary'):
                 summs.add(bs_obj.get('summary'))
         return sorted(summs)
+
+
+class CalculatedTreatmentSummary:
+    @calculated_property(condition='treatments', schema={
+        "title": "Treatment summary",
+        "description": "A summary of the treatments applied to the Biosample.",
+        "comment": "Do not submit. This is a calculated property",
+        "type": "string",
+        "notSubmittable": True,
+    })
+    def treatment_summary(self, request, treatments):
+        dur_dict = {}
+        for t in treatments:
+            t_obj = request.embed(t, '@@object?skip_calculated=true')
+            summ = ''
+            if t_obj.get('amount'):
+                amt = str(t_obj['amount'])
+                if amt.endswith('.0'):
+                    amt = amt[:-2]
+                a_units = t_obj['amount_units']
+                summ += (amt + ' ' + a_units + ' of ')
+            summ += (t_obj.get('treatment_term_name'))
+
+            if t_obj.get('duration'):
+                d = str(t_obj['duration'])
+                if d.endswith('.0'):
+                    d = d[:-2]
+                dur = pluralize(d, t_obj['duration_units'])
+            else:
+                dur = 'none'
+
+            if dur_dict.get(dur):
+                if t_obj.get('amount'):
+                    dur_dict[dur].append(summ)
+                else:
+                    dur_dict[dur].insert(0, summ)
+            else:
+                dur_dict[dur] = [summ]
+
+        ovr = []
+        for k,v in dur_dict.items():
+            temp = (' and '.join(v))
+            if k != 'none':
+                temp += (' for ' + k)
+            ovr.append(temp)
+        return ('; '.join(ovr))
